requirements.txt
requests>=2.28.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
streamlit>=1.28.0
openai>=1.0.0 